Refactor and tidy up ftrace plugin
